+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unpublished
+
+- Add a trait `factory::HwndProvider`, to bridge a hwnd provider with the swapchain creation API.
+- Add a method `create_swapchain` to create a swapchain directly from a HwndProvider.
+
 # 0.2.0
 
 - Descriptor types rework.